Phase 2: Fix no-any-return and dict-item MyPy errors
- Fixed 8 no-any-return errors in secrets_manager.py by adding explicit str() casts for conn.password and extra dict values
- Fixed 3 dict-item errors in validators.py by adding explicit int/float casts
- Errors reduced from 52 to ~38 (26.9% reduction)
- Improved type safety in credential retrieval and validation reporting

+++ src/data_quality/validators.py
@@ -305,12 +305,12 @@ class DataQualityValidator:
                         {
                             "expectation_type": "expect_column_values_to_be_between",
                             "column": col,
-                            "unexpected_count": invalid_count,
-                            "unexpected_percent": (1.0 - compliance) * 100,
+                            "unexpected_count": int(invalid_count),
+                            "unexpected_percent": float((1.0 - compliance) * 100),
                             "kwargs": {
-                                "min_value": min_val,
-                                "max_value": max_val,
-                                "mostly": mostly,
+                                "min_value": float(min_val),
+                                "max_value": float(max_val),
+                                "mostly": float(mostly),
                             },
                         }
                     )
